PhotoPreProcessor/requirements.txt
easyocr==1.7.1
Flask==3.1.0
Flask-Cors==3.0.10
matplotlib==3.9.2
numpy==2.1.3
packaging==24.1
pandas==2.2.3
pillow==10.4.0
requests==2.32.3
scipy==1.14.1
tokenizers==0.21.0
torch==2.5.1+cu124
torchaudio==2.5.1
torchvision==0.20.1
transformers==4.49.0
urllib3==2.3.0
Werkzeug==3.1.3
